+++ WalletTracker.Infrastructure/Repositories/IncomeCategoryRepository.cs
@@ -49,7 +49,12 @@ namespace WalletTracker.Infrastructure.Repositories
         public async Task DeleteById(int id)
         {
             var category = await _dbContext.IncomeCategoriesAssignedToUsers
-                .FirstAsync(c => c.Id == id);
+                .FirstOrDefaultAsync(c => c.Id == id);
+
+            if (category == null)
+            {
+                throw new InvalidOperationException("Income category with specified id doesn't exist.");
+            }
 
             _dbContext.IncomeCategoriesAssignedToUsers.Remove(category);
 
@@ -68,7 +73,16 @@ namespace WalletTracker.Infrastructure.Repositories
         }
 
         public async Task<IncomeCategoryAssignedToUser> GetById(int id)
-            => await _dbContext.IncomeCategoriesAssignedToUsers
-            .FirstAsync(c => c.Id == id);
+        {
+            var category = await _dbContext.IncomeCategoriesAssignedToUsers
+                .FirstOrDefaultAsync(c => c.Id == id);
+
+            if (category == null)
+            {
+                throw new InvalidOperationException("Income category with specified id doesn't exist.");
+            }
+
+            return category;
+        }
     }
 }
